refactor structure to remove vrap references
