+++ build.gradle
@@ -254,8 +254,9 @@ def exportedProjects= [
         ":commercetools:commercetools-sdk-java-importapi",
         ":commercetools:commercetools-sdk-java-ml",
         ":rmf:rmf-java-base",
-        ":rmf:okhttp-client3",
-        ":rmf:okhttp-client4",
+        ":commercetools:commercetools-okhttp-client3",
+        ":commercetools:commercetools-okhttp-client4",
+        ":rmf:okhttp-client",
 ]
 task alljavadoc(type: Javadoc) {
     source exportedProjects.collect { project(it).sourceSets.main.allJava }
