+++ commercetools/commercetools-http-client/build.gradle
@@ -1,4 +1,4 @@
 
 dependencies {
-    api project(":rmf:okhttp-client4")
+    api project(":commercetools:commercetools-okhttp-client4")
 }
