+++ commercetools/commercetools-okhttp-client3/src/main/java/com/commercetools/http/okhttp3/BuilderOptions.java
@@ -1,0 +1,8 @@
+package com.commercetools.http.okhttp3;
+
+import okhttp3.OkHttpClient;
+
+@FunctionalInterface
+public interface BuilderOptions {
+    OkHttpClient.Builder plus(OkHttpClient.Builder builder);
+}
