+++ commercetools/commercetools-okhttp-client3/src/main/java/com/commercetools/http/okhttp3/OkHttp3ClientSupplier.java
@@ -1,0 +1,12 @@
+package com.commercetools.http.okhttp3;
+
+import io.vrap.rmf.base.client.BaseHttpClientSupplier;
+import io.vrap.rmf.base.client.VrapHttpClient;
+
+public class OkHttp3ClientSupplier extends BaseHttpClientSupplier {
+
+    @Override
+    public VrapHttpClient get() {
+        return new CtOkHttp3Client();
+    }
+}
