+++ commercetools/commercetools-okhttp-client3/src/main/resources/META-INF/services/io.vrap.rmf.base.client.BaseHttpClientSupplier
@@ -1,0 +1,1 @@
+com.commercetools.http.okhttp3.OkHttp3ClientSupplier
