+++ commercetools/commercetools-okhttp-client4/src/main/java/com/commercetools/http/okhttp4/OkHttp4ClientSupplier.java
@@ -1,0 +1,12 @@
+package com.commercetools.http.okhttp4;
+
+import io.vrap.rmf.base.client.BaseHttpClientSupplier;
+import io.vrap.rmf.base.client.VrapHttpClient;
+
+public class OkHttp4ClientSupplier extends BaseHttpClientSupplier {
+
+    @Override
+    public VrapHttpClient get() {
+        return new CtOkHttp4Client();
+    }
+}
