+++ commercetools/commercetools-okhttp-client4/src/main/resources/META-INF/services/io.vrap.rmf.base.client.BaseHttpClientSupplier
@@ -1,0 +1,1 @@
+com.commercetools.http.okhttp4.OkHttp4ClientSupplier
