+++ rmf/okhttp-client/build.gradle
@@ -1,0 +1,7 @@
+
+dependencies {
+    api project(":rmf:rmf-java-base")
+    implementation("com.squareup.okhttp3:okhttp:4.0.1")
+
+    testImplementation "junit:junit:${versions.junit}"
+}
