+++ rmf/okhttp-client/src/main/resources/META-INF/services/io.vrap.rmf.base.client.BaseHttpClientSupplier
@@ -1,0 +1,1 @@
+io.vrap.rmf.okhttp.VrapOkHttpClientSupplier
