--- rmf/okhttp-client3/src/main/java/io/vrap/rmf/okhttp/OkHttpClientSupplier.java
@@ -1,12 +1,0 @@
-package io.vrap.rmf.okhttp;
-
-import io.vrap.rmf.base.client.BaseHttpClientSupplier;
-import io.vrap.rmf.base.client.VrapHttpClient;
-
-public class OkHttpClientSupplier extends BaseHttpClientSupplier {
-
-    @Override
-    public VrapHttpClient get() {
-        return new VrapOkHttpClient();
-    }
-}
