--- rmf/okhttp-client4/src/main/java/io/vrap/rmf/okhttp/BuilderOptions.java
@@ -1,8 +1,0 @@
-package io.vrap.rmf.okhttp;
-
-import okhttp3.OkHttpClient;
-
-@FunctionalInterface
-public interface BuilderOptions {
-    OkHttpClient.Builder plus(OkHttpClient.Builder builder);
-}
