--- rmf/okhttp-client4/src/main/resources/META-INF/services/io.vrap.rmf.base.client.BaseHttpClientSupplier
@@ -1,1 +1,0 @@
-io.vrap.rmf.okhttp.OkHttpClientSupplier
