+++ settings.gradle
@@ -1,7 +1,8 @@
 rootProject.name = 'commercetools-java-sdks'
 include 'rmf:rmf-java-base'
-include 'rmf:okhttp-client3'
-include 'rmf:okhttp-client4'
+include 'commercetools:commercetools-okhttp-client3'
+include 'commercetools:commercetools-okhttp-client4'
+include 'rmf:okhttp-client'
 include 'commercetools:commercetools-http-client'
 include 'commercetools:commercetools-sdk-java-api'
 include 'commercetools:commercetools-sdk-java-importapi'
